done WITH database setup, and app startup along with seeding

+++ migrations/20210406194614-create-wallet.js
@@ -18,9 +18,9 @@ module.exports = {
         allowNull: false,
       },
       balance: {
-        type: DataTypes.INTEGER,
+        type: DataTypes.DECIMAL(23, 2),
         allowNull: false,
-        defaultValue: 0,
+        defaultValue: 0.00,
       },
       currency: {
         type: DataTypes.STRING,
